requirements.txt
asgiref==3.5.2
attrs==22.2.0
beautifulsoup4==4.12.2
boto3==1.26.117
botocore==1.29.117
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.1
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.1.4
django-cors-headers==3.13.0
django-import-export==3.2.0
django-phonenumber-field==7.0.1
django-storages==1.13.2
djangorestframework==3.14.0
et-xmlfile==1.1.0
gunicorn==20.1.0
idna==3.4
Jinja2==3.1.2
jmespath==1.0.1
jsonfield==3.1.0
jsonschema==4.17.3
MarkupPy==1.14
MarkupSafe==2.1.2
mysql-connector-python==8.0.33
odfpy==1.4.1
openpyxl==3.1.2
phonenumbers==8.13.4
protobuf==3.20.3
pycparser==2.21
PyMySQL==1.0.2
pyrsistent==0.19.2
python-dateutil==2.8.2
python-decouple==3.6
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
s3transfer==0.6.1
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.3
tablib==3.5.0
typing_extensions==4.6.3
tzdata==2022.7
urllib3==1.26.16
xlrd==2.0.1
xlwt==1.3.0
